Finally got the basics of our transclusion functional.
This is kind of a huge fix/change, here.  We no longer host actual article content for wayPoint in the site's repo, except as partials within an included submodule.  Soon all site and rapidExpedition wiki content will be done in this way (it essentially already is, if we had any content to transclude).  We still need to transclude some new articles for wayPoint as well as to migrate the rapidExpedition article to the wiki (will be transcluding that too, of course).  I had a little bit of trouble figuring Liquid out, but I think I've got it licked, here.  We're very close to being able to add wikiTransclusion functionality to the master branch and main site, and then we should have 0.2 (after some minor cosmetic changes and perhaps some after drafting some decent reference material).